Offline commit: ragged-bottom = ##f
Test 1 to check how git behaves offline.
I have a working PC that is not connected to the internet
and whose working data is kept in sync through an USB stick.

When files are in sync and I run 'git status' on the offline PC
I get "nothing to commit".

Now I made a small change to the sources and commit it
on the offline PC.
When I sync the files and run 'git status' on the online PC
I assume I will also have a clean repository that I can push
without further action.

Test 2 will be changes that I do on the offline PC
but don't commit on this one.