Commit after usb transfer: Add headers
Second test how git behaves when moving the repo directory.
The changes in this commit have been made on my offline PC.
Then I transferred the directory to my online PC and committed.
As expected there are no problems:
Obviously git takes the complete directory as-is, so one can freely
move it around.
Although it is quite irritating to see all the changed git files in
the syncing tool (i.e. Unison)